+++ lib/cargoDoc.nix
@@ -3,16 +3,22 @@
 
 { cargoDocExtraArgs ? "--no-deps"
 , cargoExtraArgs ? "--locked"
+, preInstall ? ""
 , ...
 }@origArgs:
 let
   args = (builtins.removeAttrs origArgs [
     "cargoDocExtraArgs"
     "cargoExtraArgs"
+    "preInstall"
   ]);
 in
 mkCargoDerivation (args // {
   pnameSuffix = "-doc";
 
   buildPhaseCargoCommand = "cargoWithProfile doc ${cargoExtraArgs} ${cargoDocExtraArgs}";
+
+  preInstall = preInstall + ''
+    rm -rf target/release
+  '';
 })